+++ packages/identity-app/src/Identity.tsx
@@ -6,27 +6,65 @@ import { Container, Grid } from '@substrate/ui-components';
 
 import React from 'react';
 import { RouteComponentProps } from 'react-router-dom';
-import Tab from 'semantic-ui-react/dist/commonjs/modules/Tab';
+import Dropdown from 'semantic-ui-react/dist/commonjs/modules/Dropdown';
+import Menu from 'semantic-ui-react/dist/commonjs/collections/Menu';
 
 import { ManageAccounts } from './ManageAccounts';
 
 interface Props extends RouteComponentProps {}
 
+const accountManagementOptions = [
+  {
+    key: 'Edit',
+    text: 'Edit',
+    value: 'Edit'
+  },
+  {
+    key: 'Create',
+    text: 'Create',
+    value: 'Create'
+  },
+  {
+    key: 'Restore',
+    text: 'Restore',
+    value: 'Restore'
+  }
+];
+
+const addressManagementOptions = [
+  {
+    key: 'Lookup',
+    text: 'Lookup',
+    value: 'Lookup'
+  },
+  {
+    key: 'Add',
+    text: 'Add',
+    value: 'Add'
+  }
+];
+
 export class Identity extends React.PureComponent<Props> {
   render () {
-    const panes = [
-      { menuItem: 'Manage Accounts', render: () => <ManageAccounts {...this.props} /> }
-    ];
-
     return (
       <Container>
         <Grid>
           <Grid.Row stretched>
-            <Grid.Column width={10}>
-              <Tab panes={panes} menu={{ borderless: true }} />
-            </Grid.Column>
-            <Grid.Column width={6}>
-              nothing else
+            <Grid.Column width={12}>
+              <Menu>
+                <Dropdown
+                    placeholder='Manage Accounts'
+                    fluid
+                    selection
+                    options={accountManagementOptions}
+                  />
+                <Dropdown
+                    placeholder='Manage Addresses'
+                    fluid
+                    selection
+                    options={addressManagementOptions}
+                  />
+              </Menu>
             </Grid.Column>
           </Grid.Row>
         </Grid>
